requirements.txt
streamlit
numpy
pandas
scikit-learn
librosa
praat-parselmouth
pydub
audiorecorder
soundfile
matplotlib
